# backend/app/schemas/enums.py
from enum import StrEnum


class DisruptionStage(StrEnum):
    DETECTED = "DETECTED"
    DIAGNOSED = "DIAGNOSED"
    SOURCING = "SOURCING"
    AWAITING_APPROVAL = "AWAITING_APPROVAL"
    APPROVED = "APPROVED"
    REJECTED = "REJECTED"
    NEGOTIATING = "NEGOTIATING"
    NEGOTIATED = "NEGOTIATED"
    SETTLEMENT_PENDING = "SETTLEMENT_PENDING"
    SETTLED = "SETTLED"
    CLOSED = "CLOSED"
    FAILED = "FAILED"


class AgentName(StrEnum):
    SENTINEL = "SENTINEL"
    DIAGNOSIS = "DIAGNOSIS"
    SOURCING = "SOURCING"
    NEGOTIATION = "NEGOTIATION"
    SETTLEMENT = "SETTLEMENT"
    GOVERNANCE = "GOVERNANCE"


class AgentStatus(StrEnum):
    IDLE = "IDLE"
    RUNNING = "RUNNING"
    DONE = "DONE"
    BLOCKED = "BLOCKED"
    ERROR = "ERROR"


class DisruptionType(StrEnum):
    DELIVERY_DELAY = "DELIVERY_DELAY"
    VENDOR_UNRESPONSIVE = "VENDOR_UNRESPONSIVE"
    QUALITY_REJECTION = "QUALITY_REJECTION"
    STOCKOUT_RISK = "STOCKOUT_RISK"
    PRICE_SHOCK = "PRICE_SHOCK"
    LOGISTICS_HOLD = "LOGISTICS_HOLD"


class RootCause(StrEnum):
    CUSTOMS_HOLD = "CUSTOMS_HOLD"
    TRANSPORT_BREAKDOWN = "TRANSPORT_BREAKDOWN"
    VENDOR_CAPACITY = "VENDOR_CAPACITY"
    RAW_MATERIAL_SHORTAGE = "RAW_MATERIAL_SHORTAGE"
    QUALITY_FAILURE = "QUALITY_FAILURE"
    FINANCIAL_DISTRESS = "FINANCIAL_DISTRESS"
    UNKNOWN = "UNKNOWN"


class VerificationStatus(StrEnum):
    VERIFIED = "VERIFIED"
    UNVERIFIED = "UNVERIFIED"
    FAILED = "FAILED"
    UNAVAILABLE = "UNAVAILABLE"


class ApprovalDecision(StrEnum):
    APPROVE = "APPROVE"
    REJECT = "REJECT"
    REQUEST_OPTIONS = "REQUEST_OPTIONS"


class ApprovalStatus(StrEnum):
    PENDING = "PENDING"
    APPROVED = "APPROVED"
    REJECTED = "REJECTED"
    OPTIONS_REQUESTED = "OPTIONS_REQUESTED"


class Channel(StrEnum):
    WEB = "WEB"
    WHATSAPP = "WHATSAPP"
    VOICE = "VOICE"
    SYSTEM = "SYSTEM"


class DetectorSource(StrEnum):
    RULE_BASED = "RULE_BASED"
    TTM_FORECAST = "TTM_FORECAST"


class IntegrationStatus(StrEnum):
    LIVE = "LIVE"
    STUB = "STUB"
    UNAVAILABLE = "UNAVAILABLE"
    NOT_CONFIGURED = "NOT_CONFIGURED"


class MemorySource(StrEnum):
    SUPERMEMORY = "SUPERMEMORY"
    DB_ONLY = "DB_ONLY"
    UNAVAILABLE = "UNAVAILABLE"


class ForecastModel(StrEnum):
    TTM = "granite-timeseries-ttm-r2"
    RULE_BASED = "RULE_BASED"


class WSEventType(StrEnum):
    AGENT_STATUS_CHANGED = "AGENT_STATUS_CHANGED"
    DISRUPTION_CREATED = "DISRUPTION_CREATED"
    STAGE_CHANGED = "STAGE_CHANGED"
    EXPOSURE_COMPUTED = "EXPOSURE_COMPUTED"
    CANDIDATES_FOUND = "CANDIDATES_FOUND"
    APPROVAL_REQUESTED = "APPROVAL_REQUESTED"
    APPROVAL_DECIDED = "APPROVAL_DECIDED"
    NEGOTIATION_UPDATE = "NEGOTIATION_UPDATE"
    SETTLEMENT_STAGED = "SETTLEMENT_STAGED"
    FORECAST_ALERT = "FORECAST_ALERT"
    HEARTBEAT = "HEARTBEAT"
    # Demo phase D0 additions — emitting code doesn't exist yet for all of
    # these (see CLAUDE.md's "Demo phases" section); the frontend is building
    # against the shape now, not later.
    IMPACT_COMPUTED = "IMPACT_COMPUTED"
    PLAN_PROPOSED = "PLAN_PROPOSED"
    CALL_STARTED = "CALL_STARTED"
    CALL_TRANSCRIPT = "CALL_TRANSCRIPT"
    CALL_FIELD_EXTRACTED = "CALL_FIELD_EXTRACTED"
    CALL_ENDED = "CALL_ENDED"
    AGENT_SHEET_SYNCED = "AGENT_SHEET_SYNCED"
    INGEST_PROGRESS = "INGEST_PROGRESS"
    BRIEFING_READY = "BRIEFING_READY"
    CALL_TIMEOUT = "CALL_TIMEOUT"


class CallStatus(StrEnum):
    DIALING = "DIALING"
    CONNECTED = "CONNECTED"
    NEGOTIATING = "NEGOTIATING"
    CONFIRMED = "CONFIRMED"
    FAILED = "FAILED"
    ENDED = "ENDED"


class GraphNodeKind(StrEnum):
    VENDOR = "VENDOR"
    ITEM = "ITEM"
    LINE = "LINE"
    ORDER = "ORDER"
    PLANT = "PLANT"


class GraphNodeState(StrEnum):
    HEALTHY = "HEALTHY"
    AT_RISK = "AT_RISK"
    IMPACTED = "IMPACTED"
    SUBSTITUTED = "SUBSTITUTED"


class ScenarioKind(StrEnum):
    BACKED_OUT = "BACKED_OUT"
    PRICE_HIKE = "PRICE_HIKE"
    DELAYED = "DELAYED"
    SHUT_DOWN = "SHUT_DOWN"


class PlanChangeKind(StrEnum):
    SPLIT_ORDER = "SPLIT_ORDER"
    SWITCH_VENDOR = "SWITCH_VENDOR"
    PULL_FORWARD_STOCK = "PULL_FORWARD_STOCK"
    REDUCE_QUANTITY = "REDUCE_QUANTITY"
    EXPEDITE_FREIGHT = "EXPEDITE_FREIGHT"


class IngestStatus(StrEnum):
    QUEUED = "QUEUED"
    PARSING = "PARSING"
    RESOLVED = "RESOLVED"
    FAILED = "FAILED"