# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sanjeevani Backend",
    "description": "Sanjeevani supply-chain disruption system backend.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/disruptions/simulate": {
      "post": {
        "tags": [
          "simulate"
        ],
        "summary": "Simulate Disruption",
        "operationId": "simulate_disruption_api_v1_disruptions_simulate_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/simulate/targets": {
      "get": {
        "tags": [
          "simulate"
        ],
        "summary": "Simulate Targets",
        "operationId": "simulate_targets_api_v1_simulate_targets_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateTargetsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScenarioKind": {
        "type": "string",
        "enum": [
          "BACKED_OUT",
          "PRICE_HIKE",
          "DELAYED",
          "SHUT_DOWN"
        ],
        "title": "ScenarioKind"
      },
...
      "SimulateRequest": {
        "properties": {
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          },
          "vendor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          }
        },
        "type": "object",
        "title": "SimulateRequest"
      },
      "SimulateResponse": {
        "properties": {
          "disruption_id": {
            "type": "string",
            "title": "Disruption Id"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "newly_triggered": {
            "type": "boolean",
            "title": "Newly Triggered"
          }
        },
        "type": "object",
        "required": [
          "disruption_id",
          "scenario",
          "stage",
          "newly_triggered"
        ],
        "title": "SimulateResponse"
      },
      "SimulateTarget": {
        "properties": {
          "vendor_id": {
            "type": "string",
            "title": "Vendor Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "open_po_count": {
            "type": "integer",
            "title": "Open Po Count"
          },
          "downstream_line_count": {
            "type": "integer",
            "title": "Downstream Line Count"
          },
          "est_exposure_paise": {
            "type": "integer",
            "title": "Est Exposure Paise"
          },
          "recommended_kinds": {
            "items": {
              "$ref": "#/components/schemas/ScenarioKind"
            },
            "type": "array",
            "title": "Recommended Kinds"
          }
        },
        "type": "object",
        "required": [
          "vendor_id",
          "name",
          "category",
          "open_po_count",
          "downstream_line_count",
          "est_exposure_paise",
          "recommended_kinds"
        ],
        "title": "SimulateTarget"
      },
      "SimulateTargetsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SimulateTarget"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SimulateTargetsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}